# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIRAMA API",
        "description": "API untuk aplikasi mobile SIRAMA (Sistem Informasi Administrasi Kependudukan). Autentikasi menggunakan Laravel Sanctum Bearer Token.",
        "contact": {
            "email": "admin@sirama.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://sirama.tunggulmajid.my.id/api",
            "description": "SIRAMA API Server (Production)"
        },
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "SIRAMA API Server (Local)"
        }
    ],
    "paths": {
...
        "/pengajuan/akta-lahir": {
            "post": {
                "tags": [
                    "Pengajuan - Akta Lahir"
                ],
                "summary": "Buat pengajuan Akta Kelahiran baru",
                "description": "Membuat pengajuan Akta Kelahiran sekaligus mengisi form dan mengupload semua dokumen dalam satu request (multipart/form-data).",
                "operationId": "c3ba39078e7bd264682fb658b9b7c4be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nama_anak",
                                    "tanggal_lahir_anak",
                                    "file_sk_lahir",
                                    "file_kk",
                                    "file_ktp_ayah",
                                    "file_ktp_ibu",
                                    "file_surat_nikah"
                                ],
                                "properties": {
                                    "nama_anak": {
                                        "type": "string",
                                        "example": "Rian Santoso"
                                    },
                                    "tanggal_lahir_anak": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-10"
                                    },
                                    "file_sk_lahir": {
                                        "description": "Surat keterangan lahir dari bidan/klinik/RS/pemerintah desa",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_kk": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_ktp_ayah": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_ktp_ibu": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_surat_nikah": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengajuan Akta Lahir berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pengajuan/akta-lahir/{id}": {
            "post": {
                "tags": [
                    "Pengajuan - Akta Lahir"
                ],
                "summary": "Update pengajuan Akta Kelahiran",
                "operationId": "ac34abb126311f38ef9d6d74029dfe96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nama_lengkap": {
                                        "type": "string"
                                    },
                                    "nik": {
                                        "type": "string"
                                    },
                                    "no_whatsapp": {
                                        "type": "string"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "enum": [
                                            "L",
                                            "P"
                                        ]
                                    },
                                    "pekerjaan": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "alamat": {
                                        "type": "string"
                                    },
                                    "desa": {
                                        "type": "string"
                                    },
                                    "rt": {
                                        "type": "string"
                                    },
                                    "rw": {
                                        "type": "string"
                                    },
                                    "nama_anak": {
                                        "type": "string"
                                    },
                                    "tanggal_lahir_anak": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "file_sk_lahir": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_kk": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_ktp_ayah": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_ktp_ibu": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_surat_nikah": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pengajuan Akta Lahir berhasil diupdate"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum. Contoh: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}